requirements.txt
discord.py == 0.9.2
six == 1.10.0
python-RIBBIT == 1.0.2.dev0
frogsay == 2.0.0
